CHROMIUM: config: Enable AVS driver for APL and AML
This is a squashed commit of 81 for test purposes only.

BUG=b:219931843
TEST=AVS works on APL and AML

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3666996
Commit-Queue: Radoslaw Biernacki <biernacki@google.com>
Tested-by: Radoslaw Biernacki <biernacki@google.com>

+++ sound/soc/intel/avs/trace.c
@@ -25,7 +25,7 @@ void trace_avs_msg_payload(const void *data, size_t size)
 	while (remaining > 0) {
 		u32 chunk;
 
-		chunk = min(remaining, MAX_CHUNK_SIZE);
+		chunk = min(remaining, (size_t)MAX_CHUNK_SIZE);
 		trace_avs_ipc_msg_payload(data, chunk, offset, size);
 
 		remaining -= chunk;
